+++ lib/main_db/new/main_store_factory.dart
@@ -1,9 +1,0 @@
-import 'package:hoplixi/main_db/core/main_store.dart';
-import 'package:hoplixi/main_db/core/models/dto/main_store_dto.dart';
-
-typedef Session = ({MainStore store, StoreInfoDto info});
-
-class MainStoreFactory {
-
-  
-}